Use LCP_POLELT_TYPE_CUSTOM to wrap tboot policy, so tboot policy is allowed to be stored in the LCP index and be signed, etc. Additionally tboot code unwraps the policy from the PO index.

+++ tboot/common/loader.c
@@ -50,7 +50,6 @@
 #include <elf_defns.h>
 #include <linux_defns.h>
 #include <tb_error.h>
-#include <lcp2.h>
 #include <txt/txt.h>
 
 /* copy of kernel/VMM command line so that can append 'tboot=0x1234' */
@@ -69,7 +68,7 @@ extern bool expand_linux_image(const void *linux_image, size_t linux_size,
                                void **entry_point, bool is_measured_launch);
 extern bool jump_elf_image(void *entry_point);
 extern bool jump_linux_image(void *entry_point);
-extern bool is_sinit_acmod(void *acmod_base, uint32_t acmod_size, bool quiet);
+extern bool is_sinit_acmod(void *acmod_base, size_t acmod_size, bool quiet);
 
 #if 0
 void print_mbi(multiboot_info_t *mbi)
